End of support
This is not a project I'm proud because I did. Project's purpose was different at first and after 151 days I decided to end the project since it is using high computational resources for just downloading and it is against what Google asked for us to not to do. Thanks to @IEWbgfnYDwHRoRRSKtkdyMDUzgdwuBYgDKtDJWd for his gently warning. Fixes #15